A_r() checks permutations
A_r() now checks permutations of singularities and cyclic permutations of copies of the plane associated to a particular singularity.  However, when X is the L-surface, the matrix [[1,0],[2,1]] is not kept by A_r(), so this function still needs some work.  I've looked into this example, and the permutations of the planes are not as straight-forward as expected.

I've also sped up the computation of holonomy vectors in the marked_periods() function.  There's a predefined function for computing these in the flatsurf package that I was unaware of before.