A_r() updated
I believe the A_r() function is working for any translation surface of any stratum (though it takes a while to run).  

I edited the section of the code that computes `pre-candidate' matrices for the Veech group by considering only linearly independent holonomies corresponding to marked periods in adjacent copies of the plane that come from singularities of the same cone angle as the cone angle of the singularity corresponding to our fixed basis.

I've also rearranged a few sections of the code (e.g., the construction of permutation groups was previously inside a for loop running through matrices, but these groups only depend on the singularities of the surface so I brought that construction outside the loop).

Some of the structure of the comments has been edited as well.